+++ ecosystem/rootage/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./analyzer";
 export * from "./languages";
 export * from "./parser";
+export * from "./generator";